rename:daily -> dailry 로 네이밍 수정
* 서비스명 변경에 맞게 daily, da-ily, Daily, Da-ily 로 표기된 파일, 디렉토리, 함수, 변수 명 수정
