+++ frontend/public/index.html
@@ -7,7 +7,7 @@
       name="description"
       content="나만의 다이어리 꾸미기, 다일리"
     />
-    <title>Da-ily</title>
+    <title>Dailry</title>
   </head>
   <body>
     <div id="root"></div>
