+++ frontend/src/pages/DailryPage/DailryPage.styled.jsx
